tot/pyproject.toml
[tool.poetry]
name = "tot"
version = "0.1.0"
description = ""
authors = ["thinh.dang <thinh.dang@trustingsocial.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
tree-of-thoughts = {git = "https://github.com/kyegomez/tree-of-thoughts", rev = "0.3.6"}
langchain = "0.0.196"
transformers = "^4.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
